+++ Assets/Prefabs/Managers.prefab
@@ -13,6 +13,9 @@ GameObject:
   - component: {fileID: 5902595311601403741}
   - component: {fileID: 5902595311601403742}
   - component: {fileID: 5902595311601403740}
+  - component: {fileID: 1133189570}
+  - component: {fileID: 1133189571}
+  - component: {fileID: 1133189572}
   m_Layer: 0
   m_Name: Managers
   m_TagString: Untagged
@@ -61,24 +64,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   party: {fileID: 5902595311601403742}
   chatBroadcaster: {fileID: 5902595311601403730}
-  _playerJobActionsLists:
-  - {fileID: 11400000, guid: bd82d9fdf5b21804caff0ce69b578f01, type: 2}
-  - {fileID: 11400000, guid: 372d059ea52793146a680ded329e4347, type: 2}
-  - {fileID: 11400000, guid: d09a40123d671e8488ff6b4affb9cdb3, type: 2}
-  - {fileID: 11400000, guid: 560d63ceaa43a8146b242f0b21c603c9, type: 2}
-  _enemyJobActionsLists:
-  - {fileID: 11400000, guid: 5aeeb95751357cf4f94f3c4271ef2de5, type: 2}
-  currentStageIndex: 0
-  nextStageIndex: 0
-  stageInfo:
-  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
-  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
-  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
-  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
-  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
-  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
-  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
-  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
+  models: {fileID: 1133189570}
+  gameState: {fileID: 1133189571}
 --- !u!114 &5902595311601403742
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -123,3 +110,58 @@ MonoBehaviour:
     _speed: 0
     _hpRegen: 0
     _manaRegen: 0
+--- !u!114 &1133189570
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5902595311601403729}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1658ebc3e5a22744f885d30531896daf, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _playerJobActionsLists:
+  - {fileID: 11400000, guid: bd82d9fdf5b21804caff0ce69b578f01, type: 2}
+  - {fileID: 11400000, guid: 560d63ceaa43a8146b242f0b21c603c9, type: 2}
+  - {fileID: 11400000, guid: d09a40123d671e8488ff6b4affb9cdb3, type: 2}
+  - {fileID: 11400000, guid: 372d059ea52793146a680ded329e4347, type: 2}
+  _enemyJobActionsLists:
+  - {fileID: 11400000, guid: 5aeeb95751357cf4f94f3c4271ef2de5, type: 2}
+  stageInfo:
+  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
+  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
+  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
+  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
+  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
+  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
+  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
+  - {fileID: 11400000, guid: 6cbd221c5fb16894eae59016beb7a9ae, type: 2}
+--- !u!114 &1133189571
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5902595311601403729}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: cbff3dc10b6f94d47a278744f4e7ab1a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  progressData: {fileID: 1133189572}
+--- !u!114 &1133189572
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5902595311601403729}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6b3dd4e88a83fa741b684d20ef0c59b0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  currentStageIndex: 0
+  nextStageIndex: 0
