+++ Assets/_Scripts/GameState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cbff3dc10b6f94d47a278744f4e7ab1a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
