+++ Assets/_Scripts/ProgressData.cs
@@ -1,0 +1,10 @@
+
+using UnityEngine;
+
+// This class represents the current progress, and possibly a save state in the future
+public class ProgressData : MonoBehaviour {
+
+    public int currentStageIndex = 0;
+    public int nextStageIndex = 0;
+    
+}
