+++ Assets/_Scripts/ProgressData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6b3dd4e88a83fa741b684d20ef0c59b0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
